+++ src/js/Helper/https-debug.js
@@ -1,0 +1,5 @@
+(function() {
+    if(location.protocol !== 'https:' && location.href.indexOf('?https=false') === -1) {
+        location.href = `${location.origin}${location.pathname}?https=false`;
+    }
+}());