+++ src/vue/Details/Password/Revisions.vue
@@ -1,10 +1,14 @@
 <template>
     <ul class="revision-list">
-        <li class="revision" v-for="revision in getRevisions" :key="revision.id">
+        <li class="revision" v-for="revision in getRevisions" :key="revision.id" @click="viewAction(revision)">
             <img class="icon" :src="revision.icon" alt="">
             <span class="label">{{ revision.label }}</span>
             <span class="time">{{ getDateTime(revision.created) }}</span>
-            <translate class="restore" icon="undo" title="Restore revision" @click="restoreAction(revision)" v-if="revision.id !== password.revision"/>
+            <translate class="restore"
+                       icon="undo"
+                       title="Restore revision"
+                       @click="restoreAction(revision, $event)"
+                       v-if="revision.id !== password.revision"/>
         </li>
     </ul>
 </template>
@@ -33,8 +37,13 @@
         },
 
         methods: {
-            restoreAction(revision) {
-                PasswordManager.restoreRevision(this.password, revision);
+            restoreAction(revision, $event) {
+                $event.stopPropagation();
+                PasswordManager.restoreRevision(this.password, revision)
+                    .catch(console.error);
+            },
+            viewAction(revision) {
+                PasswordManager.viewRevision(this.password, revision);
             },
             getDateTime(date) {
                 return Localisation.formatDateTime(date);