+++ CHANGELOG.md
@@ -3,12 +3,15 @@
  - Swedish translation
  - Enpass Json Import
  - Custom fields support in export & import
+ - Password revision detail view
  - Password names in legacy api
  - Changelog in app store
 ### Changed
  - Improved Passman import
  - Improved help section
  - Improved sharing ui
+### Fixed
+ - Editable status of shares of shared password not updated
 
 ## 2019.4.2 - Bugfix release
 ### Fixed